Added "-adv" advanced options dialog, Parse GroupIDs from names
Now "MainMesh" has been renamed to "Group", and changing the Group number changes what it is saved as in the file.
The new global option "bReadGroupIds" enables/disables this feature.
The "-adv" Advanced Option will display the NoeWin plugin window on export, even when being exported from a wrapper (a checkbox "Show Advanced Options" is now in REEM for this)